requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
httpx==0.24.0
sqlalchemy==1.4.46
pydantic==1.10.7
python-multipart==0.0.6
watchdog==3.0.0
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
asyncio==3.4.3
aiohttp==3.9.1
python-logstash==0.4.8
filelock==3.13.1
# Thêm các thư viện cần thiết
